Use API to get modpack names, serialize the JSON and process it. Also show it in UI.
On initialization, the modpack list will be filled with the names of the modpacks that people can plan. This list is gathered using the REST API and processed using the Jackson Library. This will work for now, later on I will need to look into object mapping to make it more dynamic.