Implement UploadedFile model which includes information such as the download url, filename, size etc. The download url is now being added to the modpack object. This is being added to the database. This is being used in the enduser UI to download the zip. This is not being done by a rest api call yet. Just by a simple http call atm. Extraction also still needs to be done.
The EndUser UI is now getting the actual modpack object instead of a string with the modpack name. This is done so we can use the download url from the selected modpack. 
There were far more changes but I forgot all of them. Finally this painful update is out. We can now progress and finish this launcher.